# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Get information about airports",
        "title": "Airport API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3333",
    "basePath": "/api",
    "paths": {
        "/airports": {
            "get": {
                "description": "Get list of airports in the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airports"
                ],
                "summary": "Get list of airports in the database",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Airport"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/airports/{id}/averages": {
            "get": {
                "description": "Get averages of measures (temperature, pressure, wind) in a airport",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airports"
                ],
                "summary": "Get averages of measures (temperature, pressure, wind) in a airport",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of the airport (example : NTE)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "date (example : 2021-04-04)",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Average"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/airports/{id}/measures": {
            "get": {
                "description": "Get measures (temperature, pressure, wind) in a airport between two date",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airports"
                ],
                "summary": "Get measures (temperature, pressure, wind) in a airport between two date",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of the airport (example : NTE)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "sensor type (temperature, wind, pressure)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "start date (example : 2021-04-04T22:08:41Z)",
                        "name": "startDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "end date (example : 2021-04-04T22:08:41Z)",
                        "name": "endDate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Measure"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "helper.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "models.Airport": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Average": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "average": {
                    "type": "string"
                }
            }
        },
        "models.Measure": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "airportid": {
                    "type": "string"
                },
                "date": {
                    "type": "integer"
                },
                "sensorid": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        }
    }
}